feat(kikan): lift diagnostics into control_plane pure fns [#508]
Wave C of PR-B. Relocates `collect()` and bundle export into
`kikan::control_plane::diagnostics` with transport-neutral signatures:

    pub async fn collect(&PlatformState)
        -> Result<DiagnosticsResponse, ControlPlaneError>
    pub async fn build_bundle(&PlatformState)
        -> Result<(Vec<u8>, String), ControlPlaneError>

HTTP handlers in `platform::{diagnostics, diagnostics_bundle}` become
thin delegations — only the Axum response shell (Json for /diagnostics,
Content-Type + Content-Disposition headers for /diagnostics/bundle)
stays in the adapter. The log-redaction helpers (redact_patterns,
scrub_line) and their 6 unit tests moved with the pure fn — that's
their real home.

Signature takes `&PlatformState` (not `&ControlPlaneState`) because
diagnostics only reads platform fields. UDS/CLI callers holding a
`ControlPlaneState` pass `&state.platform`. Taking the narrower slice
is honest about the real dependency and keeps the HTTP mount at
`services/api::build_app_inner` (`.with_state(state.platform_state())`)
unchanged — no remount ripple into backup_status + demo reset.

Internal error seam: every failure path (DB pool read, tokio::fs,
zip build) is semantically Internal — wrapped via
`ControlPlaneError::Internal(anyhow::anyhow!(...))`. HTTP renders 500;
UDS renders identically.

The Commitment 3 DTO keys (uds_path, uds_mode, meta_db_ok, profiles[],
etc.) are deferred to PR-D where UDS + CLI callers actually consume
them; PR-B has no consumer and shipping the schema dead would drift.

Route shapes, zip archive contents, metadata JSON layout, and
redaction patterns unchanged. Existing `compute_disk_warning` re-
exported from `platform::diagnostics` so the `services/api` health
check keeps its import path.

Verified:
- cargo test -p kikan --lib ................. 215 passed
- cargo test -p kikan --test bdd ............. 50 passed, 51 skipped
- cargo test -p kikan --test control_plane_purity .. 3 passed
- cargo test -p mokumo-api --lib .............. 63 passed
- cargo clippy --workspace --all-targets ...... clean
- cargo fmt --all --check ..................... clean
- I1 / I2 / I4 invariant scripts .............. all green
