+++ CHANGELOG.md
@@ -8,6 +8,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 
 ### Internal
 
+- **Diagnostics pure-fn lift (Wave C, PR-B)**: Relocates the diagnostics snapshot collection (`collect`) and bundle export (`build_bundle`) from `kikan::platform::{diagnostics, diagnostics_bundle}` into `kikan::control_plane::diagnostics` with transport-neutral signatures (`&PlatformState → Result<_, ControlPlaneError>`). The HTTP handlers are now thin delegations that only add the Axum response shell (`Json` for `GET /api/diagnostics`, `Content-Type`/`Content-Disposition` headers for `GET /api/diagnostics/bundle`). The log-redaction helpers (`redact_patterns`, `scrub_line`) and their 6 unit tests moved with the pure fn. Signature takes `&PlatformState` (not `&ControlPlaneState`) because diagnostics only reads platform fields — UDS/CLI callers holding a `ControlPlaneState` pass `&state.platform`. Route shapes, zip archive contents, metadata JSON layout, and redaction patterns unchanged; hurl/BDD green without edits. The Commitment 3 DTO keys (`uds_path`, `uds_mode`, `meta_db_ok`, `profiles[]`, etc.) are deferred to PR-D where UDS and CLI callers actually consume them. (#508)
+- **User/auth pure-fn lift (Wave B, PR-B)**: Introduces `kikan::control_plane::users::*` (6 pure fns: `bootstrap_first_admin`, `verify_credentials`, `verify_credentials_struct`, `soft_delete_user`, `update_user_role`, `regenerate_recovery_codes`) over a new `ControlPlaneState` slice that extends `PlatformState` with the auth rate limiters (`login_limiter`, `recovery_limiter`, `regen_limiter`, `switch_limiter`, all `Arc<RateLimiter>`), `reset_pins`, `recovery_dir`, `setup_token`, `setup_in_progress`, and `activity_writer`. `kikan::platform::{auth, users}` handlers rewired as thin delegations: session issuance, rate-limit enforcement, lockout accounting, and HTTP-specific error messages ("Admin access required", "Invalid password") stay in the adapter via handler-local mappers (`map_user_admin_error`, `map_regenerate_error`); the pure-fn layer reports semantic variants only. `PendingReset` relocated from `platform::auth` to `control_plane::state`. `MokumoAppState::control_plane_state()` accessor mirrors `platform_state()`. Expands `ConflictKind` with `LastAdminProtected { message: String }` so the last-admin guard's 409 wire shape is preserved byte-for-byte with dynamic messaging; `message()` loosened to `&str`. `control_plane_error_variants.feature` Examples table refactored to one row per `ConflictKind`. Purity witness test (`crates/kikan/tests/control_plane_purity.rs`) bans `axum::/axum_login::/tower::/tower_sessions::/tower_http::/http::/mokumo_shop::` imports under `crates/kikan/src/control_plane/**`. (#508)
 - **`profile_db_middleware` consolidated into `kikan::profile_db` (Wave A.5)**: Relocates the profile-routing middleware from `services/api/src/profile_db.rs` (deleted) into `crates/kikan/src/profile_db.rs` alongside the `ProfileDb` / `ActiveProfile` extractors. The middleware is parameterized on `State<PlatformState>` — it already only touched `db_for(mode)` and `active_profile`, both present on `PlatformState` since the platform-slice lift, so the prior file-header claim that it "cannot move into platform code without dragging AppState along" was stale. Mount site in `services/api::build_app_inner` updated from `state.clone()` + `profile_db::profile_db_middleware` to `state.platform_state()` + `kikan::profile_db::profile_db_middleware`; route shapes, layer ordering, and `profile_middleware.feature` BDD contract all unchanged. (#508)
 - **UserRepo composite-method atomicity (Wave A.3b)**: Adds two new composite methods on `SeaOrmUserRepo` — `create_user_with_codes(&CreateUser, codes_count)` (generic user + recovery-codes batch with activity-log-in-txn) and `bootstrap_admin_with_codes(email, name, password)` (first-admin bootstrap with an `ALREADY_BOOTSTRAPPED` guard checked *inside* the transaction, so concurrent bootstraps race-safely). The third method named in the feature, `regenerate_codes_with_log`, is satisfied by the existing `regenerate_recovery_codes`. Introduces `kikan::auth::repo::BootstrapError` (`AlreadyBootstrapped`, `Domain(DomainError)`) with `From<BootstrapError> for ControlPlaneError` so handlers render the wire code without string-sniffing. Adds `ActivityAction::Bootstrap` (displays as `"bootstrap"`). Un-tags `user_repo_atomicity.feature` — all 6 scenarios pass (create success, validation rollback via `codes_count` out-of-range, regenerate success, regenerate log-failure rollback via a `RAISE(ABORT)` trigger on `activity_log`, bootstrap success, bootstrap `ALREADY_BOOTSTRAPPED` rejection). Adds `crates/kikan/tests/bdd_world/user_repo_atomicity_steps.rs` and `crates/kikan/tests/user_repo_trait_signature.rs` as a compile-time witness that `UserRepository` leaks no `sea_orm::DatabaseTransaction` or `sea_orm::TransactionTrait` into its public surface. Composite-method transactional atomicity is now enforced at both the structural (trait signature) and behavioral (BDD) levels. (#508)
 - **Activity HTTP handler lift into `kikan::platform::activity_http`**: Relocates `services/api/src/activity/mod.rs` (~55 LOC) to `crates/kikan/src/platform/activity_http/mod.rs`. The `GET /api/activity` list endpoint (optional `entity_type`, `entity_id`, `page`, `per_page` filters) now lives alongside the other platform handlers. Module named `activity_http` to avoid collision with `kikan::activity` (the writer/repo). `activity_router()` is generic over state (`Router<S>`) because the handler uses only per-request extractors (`ProfileDb`, `Query`). services/api mount updated to `kikan::platform::activity_http::activity_router()`; `services/api/src/activity/` deleted. Route shape and pagination semantics unchanged. (#508)
