+++ crates/kikan/src/control_plane/diagnostics.rs
@@ -1,0 +1,408 @@
+//! Transport-neutral diagnostics collection and bundle export.
+//!
+//! Lifted from `kikan::platform::{diagnostics, diagnostics_bundle}` in
+//! Wave C (PR-B). The HTTP handlers in `platform::*` are now thin
+//! delegations over these pure fns — the same entry points serve the
+//! UDS admin adapter (`kikan-admin-adapter`, PR-D) and one-shot CLI
+//! subcommands (`mokumo-server diagnose`) without re-implementing the
+//! sysinfo refresh, profile-DB inspection, or log redaction logic.
+//!
+//! ## Signature choice: `&PlatformState`, not `&ControlPlaneState`
+//!
+//! Diagnostics only reads platform fields (`data_dir`, `production_db`,
+//! `demo_db`, `mdns_status`, `active_profile`, `is_first_launch`,
+//! `started_at`). Taking the narrower slice is honest about the real
+//! dependency and lets the HTTP handler stay mounted on
+//! `PlatformState` without a remount. UDS/CLI callers holding a
+//! `ControlPlaneState` simply pass `&state.platform`.
+//!
+//! ## Error mapping seam
+//!
+//! Every internal error path (DB pool read, `tokio::fs::metadata`, zip
+//! build) is semantically `Internal` — wrap at the seam via
+//! `ControlPlaneError::Internal(anyhow::anyhow!(...))`. The HTTP
+//! adapter renders that as 500; UDS renders it identically.
+
+use std::io::{BufRead as _, BufReader, Cursor, Write as _};
+use std::path::{Path, PathBuf};
+use std::sync::OnceLock;
+
+use chrono::Utc;
+use kikan_types::diagnostics::{
+    AppDiagnostics, DatabaseDiagnostics, DiagnosticsResponse, OsDiagnostics, ProfileDbDiagnostics,
+    RuntimeDiagnostics, SystemDiagnostics,
+};
+use regex::Regex;
+use sea_orm::DatabaseConnection;
+use sysinfo::{Disks, System};
+use zip::ZipWriter;
+use zip::write::SimpleFileOptions;
+
+use crate::{ControlPlaneError, PlatformState, SetupMode};
+
+/// Collect the full diagnostics snapshot. Shared by the HTTP
+/// `GET /api/diagnostics` handler and the bundle export so sysinfo is
+/// refreshed in one place.
+pub async fn collect(state: &PlatformState) -> Result<DiagnosticsResponse, ControlPlaneError> {
+    let production_db_path = profile_db_path(&state.data_dir, SetupMode::Production);
+    let demo_db_path = profile_db_path(&state.data_dir, SetupMode::Demo);
+
+    let production = read_profile_diagnostics(&state.production_db, &production_db_path).await?;
+    let demo = read_profile_diagnostics(&state.demo_db, &demo_db_path).await?;
+
+    let mdns = state.mdns_status.read().clone();
+    let lan_url = if mdns.active {
+        mdns.hostname
+            .as_ref()
+            .map(|h| format!("http://{}:{}", h, mdns.port))
+    } else {
+        None
+    };
+    let host = mdns
+        .hostname
+        .clone()
+        .unwrap_or_else(|| mdns.bind_host.clone());
+
+    let runtime = RuntimeDiagnostics {
+        uptime_seconds: state.started_at.elapsed().as_secs(),
+        active_profile: *state.active_profile.read(),
+        setup_complete: state.is_setup_complete(),
+        is_first_launch: state
+            .is_first_launch
+            .load(std::sync::atomic::Ordering::Acquire),
+        mdns_active: mdns.active,
+        lan_url,
+        host,
+        port: mdns.port,
+    };
+
+    let system = collect_system_diagnostics(&state.data_dir);
+
+    Ok(DiagnosticsResponse {
+        app: AppDiagnostics {
+            name: env!("CARGO_PKG_NAME").into(),
+            version: env!("CARGO_PKG_VERSION").into(),
+            build_commit: option_env!("VERGEN_GIT_SHA").map(Into::into),
+        },
+        database: DatabaseDiagnostics { production, demo },
+        runtime,
+        os: OsDiagnostics {
+            family: std::env::consts::OS.into(),
+            arch: std::env::consts::ARCH.into(),
+        },
+        system,
+    })
+}
+
+/// Build the diagnostics export archive in memory. Returns
+/// `(zip_bytes, suggested_filename)`; the HTTP handler sets the
+/// `Content-Type: application/zip` and `Content-Disposition` headers.
+///
+/// Contents:
+/// - `metadata.json` — pretty-printed `DiagnosticsResponse` snapshot.
+/// - `logs/mokumo*.log` — NDJSON log files from `data_dir/logs/`,
+///   line-redacted via [`scrub_line`] before being written to the zip.
+pub async fn build_bundle(state: &PlatformState) -> Result<(Vec<u8>, String), ControlPlaneError> {
+    let diag = collect(state).await?;
+    let metadata_json = serde_json::to_string_pretty(&diag).map_err(|e| {
+        ControlPlaneError::Internal(anyhow::anyhow!("failed to serialize diagnostics: {e}"))
+    })?;
+
+    let buf = Vec::new();
+    let cursor = Cursor::new(buf);
+    let mut zip = ZipWriter::new(cursor);
+    let opts = SimpleFileOptions::default().compression_method(zip::CompressionMethod::Deflated);
+
+    zip.start_file("metadata.json", opts)
+        .map_err(|e| ControlPlaneError::Internal(anyhow::anyhow!("zip start_file: {e}")))?;
+    zip.write_all(metadata_json.as_bytes())
+        .map_err(|e| ControlPlaneError::Internal(anyhow::anyhow!("zip write metadata: {e}")))?;
+
+    let log_dir = state.data_dir.join("logs");
+    if log_dir.is_dir() {
+        let patterns = redact_patterns();
+        let mut entries: Vec<_> = std::fs::read_dir(&log_dir)
+            .map_err(|e| ControlPlaneError::Internal(anyhow::anyhow!("read log dir: {e}")))?
+            .filter_map(|entry| match entry {
+                Ok(e) => Some(e),
+                Err(e) => {
+                    tracing::warn!(error = %e, "Failed to read directory entry in log dir");
+                    None
+                }
+            })
+            .collect();
+        entries.sort_by_key(|e| e.file_name());
+
+        for entry in entries {
+            let path = entry.path();
+            let name = match path.file_name().and_then(|n| n.to_str()) {
+                Some(n) if n.starts_with("mokumo") && n.ends_with(".log") => n.to_string(),
+                _ => continue,
+            };
+
+            let file = match std::fs::File::open(&path) {
+                Ok(f) => f,
+                Err(e) => {
+                    tracing::warn!(
+                        path = %path.display(),
+                        error = %e,
+                        "Skipping log file in diagnostics bundle: could not open"
+                    );
+                    continue;
+                }
+            };
+
+            let zip_path = format!("logs/{name}");
+            zip.start_file(&zip_path, opts).map_err(|e| {
+                ControlPlaneError::Internal(anyhow::anyhow!("zip start_file logs/{name}: {e}"))
+            })?;
+
+            let reader = BufReader::new(file);
+            for line in reader.lines() {
+                let line = match line {
+                    Ok(l) => l,
+                    Err(e) => {
+                        tracing::warn!(
+                            path = %path.display(),
+                            error = %e,
+                            "Error reading line from log file; skipping remainder"
+                        );
+                        break;
+                    }
+                };
+                let scrubbed = scrub_line(&line, patterns);
+                zip.write_all(scrubbed.as_bytes()).map_err(|e| {
+                    ControlPlaneError::Internal(anyhow::anyhow!("zip write logs/{name}: {e}"))
+                })?;
+                zip.write_all(b"\n").map_err(|e| {
+                    ControlPlaneError::Internal(anyhow::anyhow!("zip write logs/{name}: {e}"))
+                })?;
+            }
+        }
+    }
+
+    let cursor = zip
+        .finish()
+        .map_err(|e| ControlPlaneError::Internal(anyhow::anyhow!("zip finish: {e}")))?;
+    let zip_bytes = cursor.into_inner();
+
+    let timestamp = Utc::now().format("%Y%m%d-%H%M%S");
+    let filename = format!("mokumo-diagnostics-{timestamp}.zip");
+
+    Ok((zip_bytes, filename))
+}
+
+/// Returns `true` when available disk space for the data directory is
+/// below the threshold.
+///
+/// Threshold is read from `MOKUMO_DISK_WARNING_THRESHOLD_BYTES` (default
+/// 500 MiB). Set to `0` to disable the warning entirely — the `u64`
+/// comparison `available < 0` is never true. Returns `false` when no
+/// disk volume can be found (not a blocking condition).
+pub fn compute_disk_warning(data_dir: &Path) -> bool {
+    let threshold: u64 = std::env::var("MOKUMO_DISK_WARNING_THRESHOLD_BYTES")
+        .ok()
+        .and_then(|v| v.parse().ok())
+        .unwrap_or(524_288_000);
+
+    let disks = Disks::new_with_refreshed_list();
+    let disk = disks
+        .iter()
+        .filter(|d| data_dir.starts_with(d.mount_point()))
+        .max_by_key(|d| d.mount_point().as_os_str().len());
+
+    disk.map(|d| d.available_space() < threshold)
+        .unwrap_or(false)
+}
+
+fn collect_system_diagnostics(data_dir: &Path) -> SystemDiagnostics {
+    let mut sys = System::new();
+    sys.refresh_memory();
+
+    let hostname = System::host_name();
+
+    let disks = Disks::new_with_refreshed_list();
+    let disk = disks
+        .iter()
+        .filter(|d| data_dir.starts_with(d.mount_point()))
+        .max_by_key(|d| d.mount_point().as_os_str().len());
+
+    if disk.is_none() {
+        tracing::warn!(
+            data_dir = %data_dir.display(),
+            "No disk volume found for data directory; disk stats will be null"
+        );
+    }
+
+    SystemDiagnostics {
+        hostname,
+        total_memory_bytes: sys.total_memory(),
+        used_memory_bytes: sys.used_memory(),
+        disk_total_bytes: disk.map(|d| d.total_space()),
+        disk_free_bytes: disk.map(|d| d.available_space()),
+        disk_warning: compute_disk_warning(data_dir),
+    }
+}
+
+fn profile_db_path(data_dir: &Path, mode: SetupMode) -> PathBuf {
+    data_dir.join(mode.as_dir_name()).join("mokumo.db")
+}
+
+async fn read_profile_diagnostics(
+    db: &DatabaseConnection,
+    db_path: &Path,
+) -> Result<ProfileDbDiagnostics, ControlPlaneError> {
+    let rt = crate::db::read_db_runtime_diagnostics(db)
+        .await
+        .map_err(|e| {
+            ControlPlaneError::Internal(anyhow::anyhow!("read_db_runtime_diagnostics failed: {e}"))
+        })?;
+    let file_size_bytes = tokio::fs::metadata(db_path).await.ok().map(|m| m.len());
+
+    let db_path_owned = db_path.to_path_buf();
+    let (wal_size_bytes, vacuum_needed) =
+        match tokio::task::spawn_blocking(move || crate::db::diagnose_database(&db_path_owned))
+            .await
+        {
+            Ok(Ok(d)) => (d.wal_size_bytes, d.vacuum_needed()),
+            Ok(Err(e)) => {
+                tracing::warn!(db = %db_path.display(), "diagnose_database failed: {e}");
+                (0, false)
+            }
+            Err(e) => {
+                tracing::warn!("spawn_blocking for diagnose_database panicked: {e}");
+                (0, false)
+            }
+        };
+
+    Ok(ProfileDbDiagnostics {
+        schema_version: rt.schema_version,
+        file_size_bytes,
+        wal_mode: rt.wal_mode,
+        wal_size_bytes,
+        vacuum_needed,
+    })
+}
+
+/// Patterns applied to each log line to redact common sensitive values.
+/// Compiled once per process lifetime via `OnceLock`.
+///
+/// Pattern design for NDJSON compatibility:
+/// - `["']?\s*[:=]\s*["']?` handles both JSON (`"password":"secret"`) and
+///   plain-text (`password=secret`, `password: secret`) key-value formats.
+/// - `[^\s,{}"']+` stops at JSON delimiters to avoid swallowing adjacent
+///   fields (`\S+` was too greedy: `{"password":"x","user":"y"}` would
+///   redact `x","user":"y"}`).
+pub(crate) fn redact_patterns() -> &'static [(Regex, &'static str)] {
+    static PATTERNS: OnceLock<Vec<(Regex, &'static str)>> = OnceLock::new();
+    PATTERNS.get_or_init(|| {
+        vec![
+            (
+                Regex::new(r"(?i)bearer\s+[A-Za-z0-9\-._~+/]+=*").unwrap(),
+                "Bearer [REDACTED]",
+            ),
+            (
+                Regex::new(r#"(?i)password\s*["']?\s*[:=]\s*["']?[^\s,{}"']+"#).unwrap(),
+                "password=[REDACTED]",
+            ),
+            (
+                Regex::new(r#"(?i)secret\s*["']?\s*[:=]\s*["']?[^\s,{}"']+"#).unwrap(),
+                "secret=[REDACTED]",
+            ),
+            (
+                Regex::new(r#"(?i)api[_-]?key\s*["']?\s*[:=]\s*["']?[^\s,{}"']+"#).unwrap(),
+                "api_key=[REDACTED]",
+            ),
+        ]
+    })
+}
+
+/// Scrub one log line. Returns a `Cow` so clean lines avoid allocation
+/// entirely.
+pub(crate) fn scrub_line<'a>(
+    line: &'a str,
+    patterns: &[(Regex, &'static str)],
+) -> std::borrow::Cow<'a, str> {
+    let mut result = std::borrow::Cow::Borrowed(line);
+    for (pattern, replacement) in patterns {
+        if let std::borrow::Cow::Owned(scrubbed) = pattern.replace_all(&result, *replacement) {
+            result = std::borrow::Cow::Owned(scrubbed);
+        }
+    }
+    result
+}
+
+#[cfg(test)]
+mod tests {
+    use super::{redact_patterns, scrub_line};
+
+    #[test]
+    fn scrubs_bearer_token() {
+        let patterns = redact_patterns();
+        let result = scrub_line("Authorization: Bearer abc.def.ghi", patterns);
+        assert!(
+            !result.contains("abc.def.ghi"),
+            "bearer token not scrubbed: {result}"
+        );
+        assert!(
+            result.contains("Bearer [REDACTED]"),
+            "expected redaction marker: {result}"
+        );
+    }
+
+    #[test]
+    fn scrubs_password_field() {
+        let patterns = redact_patterns();
+        let result = scrub_line("user login password: mysecret123", patterns);
+        assert!(
+            !result.contains("mysecret123"),
+            "password not scrubbed: {result}"
+        );
+    }
+
+    #[test]
+    fn scrubs_api_key() {
+        let patterns = redact_patterns();
+        let result = scrub_line("api_key=abc123xyz", patterns);
+        assert!(
+            !result.contains("abc123xyz"),
+            "api_key not scrubbed: {result}"
+        );
+    }
+
+    #[test]
+    fn clean_line_passes_through_unchanged() {
+        let patterns = redact_patterns();
+        let input = r#"{"level":"info","message":"order created","order_id":"ord_123"}"#;
+        let result = scrub_line(input, patterns);
+        assert_eq!(result, input, "clean line should not be modified");
+    }
+
+    #[test]
+    fn scrubs_multiple_patterns_in_one_line() {
+        let patterns = redact_patterns();
+        let result = scrub_line("secret=topsecret api_key=mykey", patterns);
+        assert!(
+            !result.contains("topsecret"),
+            "secret not scrubbed: {result}"
+        );
+        assert!(!result.contains("mykey"), "api_key not scrubbed: {result}");
+    }
+
+    /// Verify `\S+` was not used: password redaction must not swallow
+    /// adjacent JSON fields.
+    #[test]
+    fn password_redaction_does_not_over_redact_json() {
+        let patterns = redact_patterns();
+        let input = r#"{"password":"secret","user":"admin"}"#;
+        let result = scrub_line(input, patterns);
+        assert!(
+            result.contains("\"user\":\"admin\""),
+            "adjacent JSON field should not be redacted: {result}"
+        );
+        assert!(
+            !result.contains("secret"),
+            "password value should be scrubbed: {result}"
+        );
+    }
+}
