+++ crates/kikan/src/control_plane/mod.rs
@@ -20,6 +20,7 @@
 //! one-shot CLI owns tty prompts — none of which leak down into the
 //! pure-fn layer.
 
+pub mod diagnostics;
 pub mod state;
 pub mod users;
